LoadTest/requirements.txt
locust==2.29.1
pyyaml==6.0.2
